Split out feed component
